Use uint journal stream IDs
Change JournalStreamId and related journaling stream-id storage from ulong to uint to reduce per-id space usage. Update binary and JSON journal framing, manager dictionaries, API surface, and related tests/benchmarks accordingly.

+++ src/Orleans.Journaling/JournalStreamId.cs
@@ -4,4 +4,4 @@ namespace Orleans.Journaling;
 /// Identifies a state.
 /// </summary>
 /// <param name="Value">The underlying identity value.</param>
-public readonly record struct JournalStreamId(ulong Value);
+public readonly record struct JournalStreamId(uint Value);
